restructure code and configure absolute path
